\name{seewave-internal}
\alias{bartlett.w}
\alias{blackman.w}
\alias{filled.contour.modif2}
\alias{flattop.w}
\alias{hamming.w}
\alias{hanning.w}
\alias{peaks}
\alias{check.pks}
\alias{peaksign}
\alias{rectangle.w}
\alias{rev.cm.colors}
\alias{rev.gray.colors.1}
\alias{rev.gray.colors.2}
\alias{rev.heat.colors}
\alias{rev.terrain.colors}
\alias{rev.topo.colors}
\alias{soscillo}
\alias{sspectro}
\alias{spectro.colors}
\alias{stft}
\alias{temp.colors}

\title{Internal Seewave Functions}

\description{
  Internal seewave functions.
}
\note{
  These functions are not to be called by the user.
}
\keyword{internal}
