\name{seewave}

\alias{seewave-package}
\alias{seewave}
\alias{Seewave}
\alias{seawave}
\alias{Seewave}

\docType{package}

\title{Sound analysis and synthesis}

\description{seewave provides functions for analysing, manipulating, displaying, editing and synthesizing time waves (particularly sound). This package processes in particular time analysis (oscillograms and envelopes), spectral content, resonance quality factor, entropy, cross correlation and autocorrelation, zero-crossing, frequency coherence, dominant frequency, analytic signal, 2D and 3D spectrograms.}

\details{

  \tabular{ll}{
    Package: \tab seewave\cr
    Type: \tab Package\cr
    Version: \tab 1.7.0\cr
    Date: \tab 2013-07-12\cr
    License: \tab GPL version 2 or newer\cr
    Contributors : \tab Ethan C. Brown, Amandine Gasc, Stefanie LaZerte,
    Jonathan Lees, Laurent Lellouch, \cr
                  \tab Sandrine Pavoine, Luis J. Villanueva-Rivera, Zev Ross, Carl G. Witthoft\cr
    Acknowledgments: \tab Michel Baylac, Charlotte Cure, Denis Dupeyron, \cr
                     \tab Arnold Fertin, Kurt Hornik, Emiliano A. Laca, Uwe Ligges, \cr
		     \tab Emmanuel Paradis, Brian Ripley, Jesse Ross, Simon Urbanek \cr
    Webpage: \tab \url{http://rug.mnhn.fr/seewave}\cr
    Discussion group : \tab \url{http://groups.google.com/group/seewave}\cr
    Reference: \tab Sueur J, Aubin T, Simonis C - Seewave: a free
    modular tool for sound analysis and synthesis.\cr
               \tab Bioacoustics, 18: 213-226.
  }
}

\author{
  Jerome Sueur <sueur@mnhn.fr>\cr 
  Thierry Aubin\cr
  Caroline Simonis\cr

  Maintainer: Jerome Sueur <sueur@mnhn.fr>\cr
}

\keyword{dplot}
\keyword{ts}

