\name{seewave}

\alias{seewave-package}
\alias{seewave}

\docType{package}

\title{Sound analysis and synthesis}

\description{seewave provides functions for analysing, manipulating, displaying, editing and synthesizing time waves (particularly sound). This package processes in particular time analysis (oscillograms and envelopes), spectral content, resonance quality factor, entropy, cross correlation and autocorrelation, zero-crossing, frequency coherence, dominant frequency, analytic signal, 2D and 3D spectrograms.}

\details{

  \tabular{ll}{
    Package: \tab seewave\cr
    Type: \tab Package\cr
    Version: \tab 2.2.3\cr
    Date: \tab 2023-10-15\cr
    License: \tab GPL version 2 or newer\cr
    Contributors : \tab Pierre Aumond, Ethan C. Brown,\cr
                   \tab Adèle de Baudouin, \cr
                   \tab Guillaume Corbeau, Camille Desjonqueres, \cr
                   \tab Marion Depraetere, Francois Fabianek, \cr
                   \tab Amandine Gasc, Sylvain Haupert, \cr
		   \tab Eric Kasten, Laurent Lellouch,\cr
		   \tab Stefanie LaZerte, Jonathan Lees, \cr
		   \tab Jean Marchal, Thibaut Marin-Cudraz, \cr
		   \tab Andre Mikulec, Sandrine Pavoine, \cr
		   \tab David Pinaud, Luis J. Villanueva-Rivera \cr
		   \tab Zev Ross, Carl G. Witthoft, \cr
		   \tab Hristo Zhivomirov \cr
    Acknowledgments: \tab Marianna Anichini, Andrey Anikin, Michel Baylac,\cr
		     \tab Charlotte Cure, Denis Dupeyron,\cr
                     \tab Kurt Fristrup, Arnold Fertin, \cr
		     \tab Sylvain Haupert, Kurt Hornik, \cr
		     \tab Yannick Jadoul, Emiliano A. Laca,  \cr
		     \tab Uwe Ligges, Duncan Murdoch, Morgane Papin, \cr
		     \tab Emmanuel Paradis, Daniel Ridley-Ellis, \cr
		     \tab Brian Ripley, Jesse Ross, \cr
		     \tab Zev Ross, Pavel Senin, David Savage, \cr
		     \tab Arvind Sowmyan, Simon Urbanek \cr
		     \tab Maria A. Wis, George Zhang \cr
    Webpage: \tab \url{https://rug.mnhn.fr/seewave/}\cr
    Discussion group : \tab \url{https://groups.google.com/g/seewave}\cr
    Source reference: \tab Sueur J, Aubin T, Simonis C (2008) - seewave: a free
    modular tool for sound analysis and synthesis.\cr
    \tab Bioacoustics, 18: 213-226.\cr
    Book: \tab Sueur J (2018) - Sound analysis and synthesis with R. Springer.
  }
}

\author{
  Jerome Sueur <sueur@mnhn.fr>\cr 
  Thierry Aubin\cr
  Caroline Simonis\cr
  Maintainer: Jerome Sueur <sueur@mnhn.fr>\cr
}

\keyword{dplot}
\keyword{ts}