\name{seewave}

\alias{seewave-package}

\alias{seewave}

\docType{package}

\title{
Time wave analysis and graphical representation 
}

\description{
seewave provides functions for analysing, manipulating, displaying, editing and synthesizing time waves (particularly sound). This package processes in particular time analysis (oscillograms and envelopes), spectral content, resonance quality factor, entropy, cross correlation and autocorrelation, zero-crossing, frequency coherence, dominant frequency, analytic signal, 2D and 3D spectrograms.
}

\details{

\tabular{ll}{
Package: \tab seewave\cr
Type: \tab Package\cr
Version: \tab 1.4.5\cr
Date: \tab 2008-09-17\cr
License: \tab GPL version 2 or newer\cr
Contributors : \tab Jonathan Lees, Martin Maechler, Sandrine Pavoine, Luis J. Villanueva-Rivera\cr
Acknowledgments: \tab Michel Baylac, Emmanuel Paradis, Arnold Fertin, Kurt Hornik
}
}

\author{
Jerome Sueur <sueur@mnhn.fr>\cr 
Thierry Aubin <thierry.aubin@u-psud.fr>\cr
Caroline Simonis <csimonis@mnhn.fr>\cr

Maintainer: Jerome Sueur <sueur@mnhn.fr>\cr

\url{http://sueur.jerome.neuf.fr/seewave.html}
}

\keyword{dplot}
\keyword{ts}

