\name{seewave}

\alias{seewave-package}

\alias{seewave}

\docType{package}

\title{
Time wave analysis and graphical representation 
}

\description{
seewave provides functions for analysing, manipulating, displaying, editing and synthesizing time waves (particularly sound). This package processes in particular time analysis (oscillograms and envelopes), spectral content, resonance quality factor, cross correlation and autocorrelation, zero-crossing, dominant frequency, 2D and 3D spectrograms.
}

\details{

\tabular{ll}{
Package: \tab seewave\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-01-05\cr
License: \tab GPL version 2 or newer\cr
Acknowledgments: \tab Michel Baylac, Emmanuel Paradis, Martin Maechler \cr
}
}

\author{
Jérôme Sueur <jerome.sueur@ibaic.u-psud.fr>\cr 
Thierry Aubin <thierry.aubin@ibaic.u-psud.fr>\cr
Caroline Simonis-Sueur <csimonis@mnhn.fr>\cr


Maintainer: Jérôme Sueur <jerome.sueur@ibaic.u-psud.fr>}

\keyword{dplot}
\keyword{ts}

}
