\name{seewave}

\alias{seewave-package}
\alias{seewave}

\docType{package}

\title{Sound analysis and synthesis}

\description{seewave provides functions for analysing, manipulating, displaying, editing and synthesizing time waves (particularly sound). This package processes in particular time analysis (oscillograms and envelopes), spectral content, resonance quality factor, entropy, cross correlation and autocorrelation, zero-crossing, frequency coherence, dominant frequency, analytic signal, 2D and 3D spectrograms.}

\details{

  \tabular{ll}{
    Package: \tab seewave\cr
    Type: \tab Package\cr
    Version: \tab 2.0.2\cr
    Date: \tab 2015-07-28\cr
    License: \tab GPL version 2 or newer\cr
    Contributors : \tab Ethan C. Brown, Camille Desjonqueres, \cr
                   \tab Marion Depraetere, Francois Fabianek, \cr
                   \tab Amandine Gasc, Laurent Lellouch, Stefanie LaZerte, \cr
		   \tab Jonathan Lees, Jean Marchal \cr
                   \tab Sandrine Pavoine, Luis J. Villanueva-Rivera,\cr
		   \tab Zev Ross, Carl G. Witthoft, Hristo Zhivomirov \cr
    Acknowledgments: \tab Michel Baylac, Charlotte Cure, Denis Dupeyron,\cr
                     \tab Krut Fristrup, Arnold Fertin, Kurt Hornik, \cr  
                     \tab Eric Kasten, Emiliano A. Laca, Uwe Ligges, \cr
		     \tab Emmanuel Paradis, Brian Ripley, Jesse Ross, Simon Urbanek \cr
    Webpage: \tab \url{http://rug.mnhn.fr/seewave}\cr
    Discussion group : \tab \url{http://groups.google.com/group/seewave}\cr
    Reference: \tab Sueur J, Aubin T, Simonis C - Seewave: a free
    modular tool for sound analysis and synthesis.\cr
               \tab Bioacoustics, 18: 213-226.\cr
  }
}

\author{
  Jerome Sueur <sueur@mnhn.fr>\cr 
  Thierry Aubin\cr
  Caroline Simonis\cr
  Maintainer: Jerome Sueur <sueur@mnhn.fr>\cr
}

\keyword{dplot}
\keyword{ts}

