\name{sheep}

\alias{sheep}

\docType{data}

\title{Sheep bleat}

\description{Recording of a sheep bleat.}

\usage{data(sheep)}

\format{A Wave object.}

\details{
  Duration = 2.47 s. Sampling frequency = 8000 hz.
}

\source{Recording by Frederic Sebe.}

\examples{
data(sheep)
oscillo(sheep,f=8000)
}

\keyword{datasets}
